Workflow commits
Forth commit and the last one of it's type:
It is about finishing "popupFunctionality.js", adding code to "background-script.js", and recreating the "BeautifySubpageFunctionality.js".

In "popupFunctionality.js" i finally finished the logic recreation ( it is finished, but a little by little if there would apear new buttons, switches, then the code would need to grow a bit ). When opening the popup it checks if "Dream Girlfriend" game tab exists. If it exists then we show all functional layout on the tab, if not then we show an message. It ( on this state of things ) is setting "beautify" switch acordingly to it's statement stored in "background-script.js" local storage. The "beautify" switch got on change event that based on it's curent state ( activated or not by user ) change it state stored in "background-script.js" local storage, and allso acordingly to "footherswitchvar, headerswitchvar,inforswitchvar,campaignsswitchvar" from "background-script.js"  local sotrage is hiding or not elements on the site page.

In "BeautifySubpageFunctionality.js", i did almostly the same as from "popupFunctionality.js". Like showing from start the layout, changing the layout of the game page acordingly to the "beautifyswitchvar" stored in "background-script.js" local storage state, and setting it's curent state ( with onchange event ) in the "background-script.js" local storage.

And in the "background-script.js" instread of functions: "SetBeoutifyCheckboxVar" and "GetOthersCheckboxVar", i created an global getter and setter of vars. Function "handleMessage" triggered by "browser.runtime.onMessage.addListener", now have 3 types of requests that it can handle: "CheckIfDGTabIsAvailable" - it checks if "Dream Girlfriend" game tab is opened
"GetVars" - it is getting the state of vars from it local storage and send as a respons them
 "SetVars" - it is setting the state of vars from it local sotrage